+++ domain/onboarding/src/main/java/com/amalitech/onboarding/preferences/OnboardingSharedPreferences.kt
@@ -1,0 +1,7 @@
+package com.amalitech.onboarding.preferences
+
+interface OnboardingSharedPreferences {
+
+    fun saveShouldShowOnboarding(shouldShow: Boolean)
+    fun loadShouldShowOnboarding(): Boolean
+}
