--- domain/onboarding/src/main/java/com/amalitech/onboarding_domain/preferences/OnboardingSharedPreferences.kt
@@ -1,7 +1,0 @@
-package com.amalitech.onboarding_domain.preferences
-
-interface OnboardingSharedPreferences {
-
-    fun saveShouldShowOnboarding(shouldShow: Boolean)
-    fun loadShouldShowOnboarding(): Boolean
-}
