+++ ui/onboarding/src/main/java/com/amalitech/onboarding/components/AuthenticationTextField.kt
@@ -1,0 +1,157 @@
+package com.amalitech.onboarding.components
+
+import android.view.KeyEvent.ACTION_DOWN
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusDirection
+import androidx.compose.ui.focus.FocusManager
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.key.Key
+import androidx.compose.ui.input.key.key
+import androidx.compose.ui.input.key.onPreviewKeyEvent
+import androidx.compose.ui.platform.LocalFocusManager
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.amalitech.core_ui.theme.LocalSpacing
+import com.amalitech.ui.onboarding.R
+import com.amalitech.core_ui.theme.BookMeetingRoomTheme
+
+@OptIn(ExperimentalComposeUiApi::class)
+@Composable
+fun AuthenticationTextField(
+    placeholder: String,
+    value: String,
+    onValueChange: (String) -> Unit,
+    modifier: Modifier = Modifier,
+    isPassword: Boolean = false,
+    textStyle: TextStyle = TextStyle(),
+    placeholderTextStyle: TextStyle = TextStyle(
+        color = MaterialTheme.colorScheme.onBackground.copy(alpha = 0.5f)
+    ),
+    keyboardOptions: KeyboardOptions = if (isPassword) KeyboardOptions(
+        imeAction = ImeAction.Next,
+        keyboardType = KeyboardType.Password
+    ) else KeyboardOptions(
+        imeAction = ImeAction.Next
+    ),
+    focusManager: FocusManager = LocalFocusManager.current,
+    onGo: () -> Unit = {},
+    keyboardActions: KeyboardActions = KeyboardActions(
+        onNext = { focusManager.moveFocus(FocusDirection.Down) },
+        onDone = { focusManager.clearFocus() },
+        onGo = { onGo() }
+    )
+) {
+    val spacing = LocalSpacing.current
+    var passwordVisible by rememberSaveable {
+        mutableStateOf(false)
+    }
+
+    TextField(
+        value = value,
+        onValueChange = onValueChange,
+        textStyle = textStyle,
+        singleLine = true,
+        visualTransformation = if (isPassword && !passwordVisible)
+            PasswordVisualTransformation()
+        else
+            VisualTransformation.None,
+        trailingIcon = {
+            if (isPassword) {
+                val image = if (passwordVisible)
+                    R.drawable.baseline_visibility_off_24
+                else
+                    R.drawable.baseline_visibility_24
+                val description = if (passwordVisible)
+                    stringResource(id = R.string.hide_password)
+                else
+                    stringResource(id = R.string.show_password)
+
+                IconButton(onClick = { passwordVisible = !passwordVisible }) {
+                    Icon(
+                        painter = painterResource(id = image),
+                        contentDescription = description
+                    )
+                }
+            }
+        },
+        modifier = modifier
+            .padding(1.dp)
+            .border(
+                width = 1.dp,
+                color = MaterialTheme.colorScheme.onBackground.copy(alpha = 0.2f),
+                shape = RoundedCornerShape(spacing.spaceExtraSmall)
+            )
+            .padding(spacing.spaceExtraSmall)
+            .onPreviewKeyEvent {
+                if (it.key == Key.Tab && it.nativeKeyEvent.action == ACTION_DOWN) {
+                    focusManager.moveFocus(FocusDirection.Down)
+                    true
+                } else if (it.key == Key.Enter) {
+                    onGo()
+                    true
+                } else {
+                    false
+                }
+            },
+        shape = RoundedCornerShape(spacing.spaceExtraSmall),
+        colors = TextFieldDefaults.colors(
+            focusedIndicatorColor = Color.Transparent,
+            unfocusedIndicatorColor = Color.Transparent,
+            focusedTextColor = MaterialTheme.colorScheme.onBackground,
+            unfocusedContainerColor = MaterialTheme.colorScheme.background,
+            unfocusedTextColor = MaterialTheme.colorScheme.onBackground,
+            focusedContainerColor = MaterialTheme.colorScheme.background,
+        ),
+        placeholder = {
+            Text(
+                text = placeholder,
+                style = placeholderTextStyle
+            )
+        },
+        keyboardOptions = keyboardOptions,
+        keyboardActions = keyboardActions
+    )
+}
+
+@Preview
+@Composable
+fun Prev() {
+    BookMeetingRoomTheme {
+        var value by rememberSaveable {
+            mutableStateOf("")
+        }
+        AuthenticationTextField(
+            placeholder = "password",
+            value = value,
+            onValueChange = { value = it },
+            modifier = Modifier.fillMaxWidth(),
+            isPassword = false,
+            focusManager = LocalFocusManager.current
+        )
+    }
+}
