--- ui/onboarding/src/main/java/com/amalitech/onboarding_presentation/login/LoginUiState.kt
@@ -1,11 +1,0 @@
-package com.amalitech.onboarding_presentation.login
-
-import com.amalitech.core.util.UiText
-
-data class LoginUiState(
-    val email: String = "",
-    val password: String = "",
-    val error: UiText? = null,
-    val finishedLoggingIn: Boolean = false,
-    val snackBarValue: UiText? = null
-)
